requirements.txt
opencv-python>=4.8
fastapi>=0.110
uvicorn[standard]>=0.27
websockets>=12.0
ultralytics>=8.3
deepface>=0.0.93
aiosqlite>=0.20
sqlalchemy[asyncio]>=2.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt==4.0.1
alembic>=1.13
